requirements.txt
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
dateparser==1.0.0
google-api-core==1.26.3
google-api-python-client==2.1.0
google-auth==1.28.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
greenlet==1.0.0
httplib2==0.19.1
idna==2.10
importlib-metadata==3.10.0
oauthlib==3.1.0
packaging==20.9
protobuf==3.15.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyodbc==4.0.30
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
SQLAlchemy==1.4.5
typing-extensions==3.7.4.3
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.4
zipp==3.4.1
